+++ app/include/view/player_setting.hpp
@@ -31,4 +31,8 @@ private:
     BRLS_BIND(brls::BooleanCell, btnBottomBar, "setting/player/bottom_bar");
     BRLS_BIND(brls::BooleanCell, btnOSDOnToggle, "setting/player/osd_on_toggle");
     BRLS_BIND(brls::BooleanCell, btnFullscreen, "setting/fullscreen");
+    BRLS_BIND(brls::BooleanCell, btnAlwaysOnTop, "setting/always_on_top");
+    BRLS_BIND(brls::SelectorCell, btnVideoMirror, "setting/video/mirror");
+    BRLS_BIND(brls::SelectorCell, btnVideoAspect, "setting/video/aspect");
+    BRLS_BIND(brls::SliderCell, btnSubsync, "setting/video/subsync");
 };