+++ app/src/utils/config.cpp
@@ -38,28 +38,38 @@ std::unordered_map<AppConfig::Item, AppConfig::Option> AppConfig::settingMap = {
     {KEYMAP, {"keymap", {"xbox", "ps", "keyboard"}}},
     {TRANSCODEC, {"transcodec", {"h264", "hevc", "av1"}}},
     {FORCE_DIRECTPLAY, {"force_directplay"}},
-    {VIDEO_QUALITY, {"video_quality",
-                        {"Auto", "1080p - 60Mbps", "1080p - 40Mbps", "1080p - 20Mbps", "720p - 8Mbps", "720p - 6Mbps",
-                            "480p - 3Mbps", "480P - 1Mbps"},
-                        {0, 60, 40, 20, 8, 6, 3, 1}}},
+    {VIDEO_QUALITY,
+        {
+            "video_quality",
+            {"Auto", "1080p - 60Mbps", "1080p - 40Mbps", "1080p - 20Mbps", "720p - 8Mbps", "720p - 6Mbps",
+                "480p - 3Mbps", "480P - 1Mbps"},
+            {0, 60, 40, 20, 8, 6, 3, 1},
+        }},
     {FULLSCREEN, {"fullscreen"}},
     {OSD_ON_TOGGLE, {"osd_on_toggle"}},
     {TOUCH_GESTURE, {"touch_gesture"}},
-    {CLIP_POINT,{"clip_point"}},
+    {CLIP_POINT, {"clip_point"}},
     {OVERCLOCK, {"overclock"}},
     {PLAYER_BOTTOM_BAR, {"player_bottom_bar"}},
     {PLAYER_SEEKING_STEP, {"player_seeking_step", {"5", "10", "15", "30"}, {5, 10, 15, 30}}},
     {PLAYER_LOW_QUALITY, {"player_low_quality"}},
-    {PLAYER_INMEMORY_CACHE, {"player_inmemory_cache", {"0MB", "10MB", "20MB", "50MB", "100MB", "200MB", "500MB"},
-                                {0, 10, 20, 50, 100, 200, 500}}},
+    {PLAYER_INMEMORY_CACHE,
+        {
+            "player_inmemory_cache",
+            {"0MB", "10MB", "20MB", "50MB", "100MB", "200MB", "500MB"},
+            {0, 10, 20, 50, 100, 200, 500},
+        }},
     {PLAYER_HWDEC, {"player_hwdec"}},
     {PLAYER_HWDEC_CUSTOM, {"player_hwdec_custom"}},
+    {PLAYER_ASPECT, {"player_aspect", {"auto", "stretch", "crop", "4:3", "16:9"}}},
+    {ALWAYS_ON_TOP, {"always_on_top"}},
+    {SINGLE, {"single"}},
     {TEXTURE_CACHE_NUM, {"texture_cache_num"}},
     {REQUEST_THREADS, {"request_threads", {"1", "2", "4", "8"}, {1, 2, 4, 8}}},
     {REQUEST_TIMEOUT, {"request_timeout", {"1000", "2000", "3000", "5000"}, {1000, 2000, 3000, 5000}}},
     {HTTP_PROXY_STATUS, {"http_proxy_status"}},
     {HTTP_PROXY_HOST, {"http_proxy_host"}},
-    {HTTP_PROXY_PORT, {"http_proxy_port"}}
+    {HTTP_PROXY_PORT, {"http_proxy_port"}},
 };
 
 static std::string generateDeviceId() {
@@ -141,6 +151,9 @@ void AppConfig::init() {
 
     // 初始化自定义的硬件加速方案
     MPVCore::PLAYER_HWDEC_METHOD = this->getItem(PLAYER_HWDEC_CUSTOM, MPVCore::PLAYER_HWDEC_METHOD);
+
+    // 初始化视频比例
+    MPVCore::VIDEO_ASPECT = this->getItem(PLAYER_ASPECT, MPVCore::VIDEO_ASPECT);
 
     // 初始化 deviceId
     if (this->device.empty()) this->device = generateDeviceId();